old-tests/submission/Infraware/Forms/contents/Forms/output_htmlFor.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>output_htmlFor</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
	 <output name=o id='output_id'></output>
	</form>
	<script>

		var output = document.getElementById("output_id");

		if (typeof(output.htmlFor) == "object") {
			test(function() {
				assert_equals(output.htmlFor.value, "", "htmlFor attribute is not exist.");
			});
		} else {
			test(function() {
				assert_unreached("htmlFor attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
